requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.2
django-multiselectfield==0.1.12
djangorestframework==3.12.4
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.14
sqlparse==0.4.1
urllib3==1.26.4
